Unify interfaces & tests for Reader & LiveReader.
