+++ wal/wal_test.go
@@ -16,447 +16,14 @@ package wal
 
 import (
 	"bytes"
-	"encoding/binary"
-	"fmt"
-	"hash/crc32"
-	"io"
 	"io/ioutil"
 	"math/rand"
 	"os"
-	"path"
-	"path/filepath"
-	"sync"
 	"testing"
-	"time"
 
 	"github.com/prometheus/tsdb/testutil"
 )
 
-type record struct {
-	t recType
-	b []byte
-}
-
-var data = make([]byte, 100000)
-var testReaderCases = []struct {
-	t    []record
-	exp  [][]byte
-	fail bool
-}{
-	// Sequence of valid records.
-	{
-		t: []record{
-			{recFull, data[0:200]},
-			{recFirst, data[200:300]},
-			{recLast, data[300:400]},
-			{recFirst, data[400:800]},
-			{recMiddle, data[800:900]},
-			{recPageTerm, make([]byte, pageSize-900-recordHeaderSize*5-1)}, // exactly lines up with page boundary.
-			{recLast, data[900:900]},
-			{recFirst, data[900:1000]},
-			{recMiddle, data[1000:1200]},
-			{recMiddle, data[1200:30000]},
-			{recMiddle, data[30000:30001]},
-			{recMiddle, data[30001:30001]},
-			{recLast, data[30001:32000]},
-		},
-		exp: [][]byte{
-			data[0:200],
-			data[200:400],
-			data[400:900],
-			data[900:32000],
-		},
-	},
-	// Exactly at the limit of one page minus the header size
-	{
-		t: []record{
-			{recFull, data[0 : pageSize-recordHeaderSize]},
-		},
-		exp: [][]byte{
-			data[:pageSize-recordHeaderSize],
-		},
-	},
-	// More than a full page, this exceeds our buffer and can never happen
-	// when written by the WAL.
-	{
-		t: []record{
-			{recFull, data[0 : pageSize+1]},
-		},
-		fail: true,
-	},
-	// Invalid orders of record types.
-	{
-		t:    []record{{recMiddle, data[:200]}},
-		fail: true,
-	},
-	{
-		t:    []record{{recLast, data[:200]}},
-		fail: true,
-	},
-	{
-		t: []record{
-			{recFirst, data[:200]},
-			{recFull, data[200:400]},
-		},
-		fail: true,
-	},
-	{
-		t: []record{
-			{recFirst, data[:100]},
-			{recMiddle, data[100:200]},
-			{recFull, data[200:400]},
-		},
-		fail: true,
-	},
-	// Non-zero data after page termination.
-	{
-		t: []record{
-			{recFull, data[:100]},
-			{recPageTerm, append(make([]byte, 1000), 1)},
-		},
-		exp:  [][]byte{data[:100]},
-		fail: true,
-	},
-}
-
-func encodedRecord(t recType, b []byte) []byte {
-	if t == recPageTerm {
-		return append([]byte{0}, b...)
-	}
-	r := make([]byte, recordHeaderSize)
-	r[0] = byte(t)
-	binary.BigEndian.PutUint16(r[1:], uint16(len(b)))
-	binary.BigEndian.PutUint32(r[3:], crc32.Checksum(b, castagnoliTable))
-	return append(r, b...)
-}
-
-// TestReader feeds the reader a stream of encoded records with different types.
-func TestReader(t *testing.T) {
-	for i, c := range testReaderCases {
-		t.Logf("test %d", i)
-
-		var buf []byte
-		for _, r := range c.t {
-			buf = append(buf, encodedRecord(r.t, r.b)...)
-		}
-		r := NewReader(bytes.NewReader(buf))
-
-		for j := 0; r.Next(); j++ {
-			t.Logf("record %d", j)
-			rec := r.Record()
-
-			if j >= len(c.exp) {
-				t.Fatal("received more records than inserted")
-			}
-			testutil.Equals(t, c.exp[j], rec, "Bytes within record did not match expected Bytes")
-		}
-		if !c.fail && r.Err() != nil {
-			t.Fatalf("unexpected error: %s", r.Err())
-		}
-		if c.fail && r.Err() == nil {
-			t.Fatalf("expected error but got none")
-		}
-	}
-}
-
-func TestReader_Live(t *testing.T) {
-	for i, c := range testReaderCases {
-		t.Logf("test %d", i)
-		dir, err := ioutil.TempDir("", fmt.Sprintf("live_reader_%d", i))
-		t.Logf("created dir %s", dir)
-		testutil.Ok(t, err)
-		defer os.RemoveAll(dir)
-
-		// we're never going to have more than a single segment file per test case right now
-		f, err := os.Create(path.Join(dir, "00000000"))
-		testutil.Ok(t, err)
-
-		// live reader doesn't work on readers created from bytes buffers,
-		// since we need to be able to write more data to the thing we're
-		// reading from after the reader has been created
-		wg := sync.WaitGroup{}
-		// make sure the reader doesn't start until at least one record is written
-		wg.Add(1)
-		go func() {
-			for i, rec := range c.t {
-				rec := encodedRecord(rec.t, rec.b)
-				n, err := f.Write(rec)
-				testutil.Ok(t, err)
-				testutil.Assert(t, n > 0, "no bytes were written to wal")
-				if i == 0 {
-					wg.Done()
-				}
-			}
-		}()
-		sr, err := OpenReadSegment(SegmentName(dir, 0))
-		testutil.Ok(t, err)
-		lr := NewLiveReader(sr)
-		j := 0
-		wg.Wait()
-	caseLoop:
-		for {
-			for ; lr.Next(); j++ {
-				rec := lr.Record()
-				t.Log("j: ", j)
-				testutil.Equals(t, c.exp[j], rec, "Bytes within record did not match expected Bytes")
-				if j == len(c.exp)-1 {
-					break caseLoop
-				}
-
-			}
-
-			// Because reads and writes are happening concurrently, unless we get an error we should
-			// attempt to read records again.
-			if j == 0 && lr.Err() == nil {
-				continue
-			}
-
-			if !c.fail && lr.Err() != nil {
-				t.Fatalf("unexpected error: %s", lr.Err())
-			}
-			if c.fail && lr.Err() == nil {
-				t.Fatalf("expected error but got none:\n\tinput: %+v", c.t)
-			}
-			if lr.Err() != nil {
-				t.Log("err: ", lr.Err())
-				break
-			}
-		}
-	}
-}
-
-func TestLiveReaderCorrupt_ShortFile(t *testing.T) {
-	// Write a corrupt WAL segment, there is one record of pageSize in length,
-	// but the segment is only half written.
-	dir, err := ioutil.TempDir("", "wal_live_corrupt")
-	testutil.Ok(t, err)
-	defer os.RemoveAll(dir)
-
-	w, err := NewSize(nil, nil, dir, pageSize)
-	testutil.Ok(t, err)
-
-	rec := make([]byte, pageSize-recordHeaderSize)
-	_, err = rand.Read(rec)
-	testutil.Ok(t, err)
-
-	err = w.Log(rec)
-	testutil.Ok(t, err)
-
-	err = w.Close()
-	testutil.Ok(t, err)
-
-	segmentFile, err := os.OpenFile(filepath.Join(dir, "00000000"), os.O_RDWR, 0666)
-	testutil.Ok(t, err)
-
-	err = segmentFile.Truncate(pageSize / 2)
-	testutil.Ok(t, err)
-
-	err = segmentFile.Close()
-	testutil.Ok(t, err)
-
-	// Try and LiveReader it.
-	m, _, err := w.Segments()
-	testutil.Ok(t, err)
-
-	seg, err := OpenReadSegment(SegmentName(dir, m))
-	testutil.Ok(t, err)
-
-	r := NewLiveReader(seg)
-	testutil.Assert(t, r.Next() == false, "expected no records")
-	testutil.Assert(t, r.Err() == io.EOF, "expected error, got: %v", r.Err())
-}
-
-func TestLiveReaderCorrupt_ReallyMessedUp(t *testing.T) {
-	// Write a corrupt WAL segment, the last record is too long for the page,
-	// but in general not too long.
-	dir, err := ioutil.TempDir("", "wal_live_corrupt")
-	testutil.Ok(t, err)
-	defer os.RemoveAll(dir)
-
-	segmentFile, err := os.OpenFile(filepath.Join(dir, "00000000"), os.O_CREATE|os.O_RDWR, 0666)
-	testutil.Ok(t, err)
-
-	writeRecord := func(len int) {
-		rec := make([]byte, len)
-		_, err = rand.Read(rec)
-		testutil.Ok(t, err)
-
-		buf := make([]byte, recordHeaderSize)
-		buf[0] = byte(recFull)
-		binary.BigEndian.PutUint16(buf[1:], uint16(len))
-		binary.BigEndian.PutUint32(buf[3:], uint32(crc32.Checksum(rec, castagnoliTable)))
-		_, err = segmentFile.Write(buf)
-		testutil.Ok(t, err)
-
-		_, err = segmentFile.Write(rec)
-		testutil.Ok(t, err)
-	}
-
-	writeRecord(pageSize / 2)
-	writeRecord(pageSize / 2)
-
-	err = segmentFile.Close()
-	testutil.Ok(t, err)
-
-	// Try and LiveReader it.
-	w, err := NewSize(nil, nil, dir, pageSize)
-	testutil.Ok(t, err)
-
-	m, _, err := w.Segments()
-	testutil.Ok(t, err)
-
-	seg, err := OpenReadSegment(SegmentName(dir, m))
-	testutil.Ok(t, err)
-
-	r := NewLiveReader(seg)
-	testutil.Assert(t, r.Next() == true, "expected one records")
-	testutil.Assert(t, r.Next() == false, "expected one records")
-	testutil.Assert(t, r.Err().Error() == "record too big for page", "expected error, got: %v", r.Err())
-}
-
-func TestLiveReaderCorrupt_RecordTooLongAndShort(t *testing.T) {
-	// Write a corrupt WAL segment, when record len > page size.
-	dir, err := ioutil.TempDir("", "wal_live_corrupt")
-	testutil.Ok(t, err)
-	defer os.RemoveAll(dir)
-
-	w, err := NewSize(nil, nil, dir, pageSize*2)
-	testutil.Ok(t, err)
-
-	rec := make([]byte, pageSize-recordHeaderSize)
-	_, err = rand.Read(rec)
-	testutil.Ok(t, err)
-
-	err = w.Log(rec)
-	testutil.Ok(t, err)
-
-	err = w.Close()
-	testutil.Ok(t, err)
-
-	segmentFile, err := os.OpenFile(filepath.Join(dir, "00000000"), os.O_RDWR, 0666)
-	testutil.Ok(t, err)
-
-	// Override the record length
-	buf := make([]byte, 3)
-	buf[0] = byte(recFull)
-	binary.BigEndian.PutUint16(buf[1:], 0xFFFF)
-	_, err = segmentFile.WriteAt(buf, 0)
-	testutil.Ok(t, err)
-
-	err = segmentFile.Close()
-	testutil.Ok(t, err)
-
-	// Try and LiveReader it.
-	m, _, err := w.Segments()
-	testutil.Ok(t, err)
-
-	seg, err := OpenReadSegment(SegmentName(dir, m))
-	testutil.Ok(t, err)
-
-	r := NewLiveReader(seg)
-	testutil.Assert(t, r.Next() == false, "expected no records")
-	testutil.Assert(t, r.Err().Error() == "record too big for page", "expected error, got: %v", r.Err())
-}
-
-func TestWAL_FuzzWriteRead_Live(t *testing.T) {
-	dir, err := ioutil.TempDir("", "wal_fuzz_live")
-	testutil.Ok(t, err)
-	defer os.RemoveAll(dir)
-
-	w, err := NewSize(nil, nil, dir, 128*pageSize)
-	testutil.Ok(t, err)
-
-	// In the background, generate a stream of random records and write them
-	// to the WAL.
-	const count = 500
-	input := make(chan []byte, count) // buffering required as we sometimes batch WAL writes.
-	done := make(chan struct{})
-	go func() {
-		var recs [][]byte
-		for i := 0; i < count; i++ {
-			var sz int64
-			switch i % 5 {
-			case 0, 1:
-				sz = 50
-			case 2, 3:
-				sz = pageSize
-			default:
-				sz = pageSize * 8
-			}
-
-			rec := make([]byte, rand.Int63n(sz))
-			_, err := rand.Read(rec)
-			testutil.Ok(t, err)
-
-			input <- rec
-
-			// Randomly batch up records.
-			recs = append(recs, rec)
-			if rand.Intn(4) < 3 {
-				testutil.Ok(t, w.Log(recs...))
-				recs = recs[:0]
-			}
-		}
-		testutil.Ok(t, w.Log(recs...))
-		time.Sleep(100 * time.Millisecond)
-		close(done)
-	}()
-
-	// Tail the WAL and compare the results.
-	m, _, err := w.Segments()
-	testutil.Ok(t, err)
-
-	seg, err := OpenReadSegment(SegmentName(dir, m))
-	testutil.Ok(t, err)
-
-	r := NewLiveReader(seg)
-	segmentTicker := time.NewTicker(100 * time.Millisecond)
-	readTicker := time.NewTicker(10 * time.Millisecond)
-
-	readSegment := func(r *LiveReader) bool {
-		for r.Next() {
-			rec := r.Record()
-			expected, ok := <-input
-			testutil.Assert(t, ok, "unexpected record")
-			testutil.Assert(t, bytes.Equal(expected, rec), "record does not match expected")
-		}
-		testutil.Assert(t, r.Err() == io.EOF, "expected EOF, got: %v", r.Err())
-		return true
-	}
-
-outer:
-	for {
-		select {
-		case <-segmentTicker.C:
-			// check if new segments exist
-			_, last, err := w.Segments()
-			testutil.Ok(t, err)
-			if last <= seg.i {
-				continue
-			}
-
-			// read to end of segment.
-			readSegment(r)
-
-			fi, err := os.Stat(SegmentName(dir, seg.i))
-			testutil.Ok(t, err)
-			testutil.Assert(t, r.TotalRead() == fi.Size(), "expected to have read whole segment, but read %d of %d", r.TotalRead(), fi.Size())
-
-			seg, err = OpenReadSegment(SegmentName(dir, seg.i+1))
-			testutil.Ok(t, err)
-			r = NewLiveReader(seg)
-
-		case <-readTicker.C:
-			readSegment(r)
-
-		case <-done:
-			readSegment(r)
-			break outer
-		}
-	}
-
-	testutil.Assert(t, r.Err() == io.EOF, "expected EOF")
-}
 func TestWAL_FuzzWriteRead(t *testing.T) {
 	const count = 25000
